requirements.txt
PyQt6>=6.6.0
python-vlc>=3.0.0
pysubs2>=1.6.0
google-genai>=0.2.0
openai>=1.0.0
anthropic>=0.18.0
requests>=2.31.0
anyio>=3.7.1,<4.0.0
cryptography>=41.0.0
openpyxl>=3.1.0
charset-normalizer>=3.3.0
langdetect>=1.0.9
faster-whisper>=1.0.0
whisperx>=3.1.1
numpy>=1.24.0
python-mpv>=1.0.8